Incorporation des Menus, premiere division de la page commande
