+++ lib/views/caisse/prisecommande/widget/caisse_produit_list_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:la_bonne_franquette_front/models/interface/identifiable.dart';
 import 'package:la_bonne_franquette_front/models/produit.dart';
-import 'package:la_bonne_franquette_front/views/caisse/widget/element_button.dart';
+import 'package:la_bonne_franquette_front/views/caisse/prisecommande/widget/element_button.dart';
 import '../viewmodel/caisse_view_model.dart';
 
 class CaisseProduitListView extends StatelessWidget {
